refactor(swc/plugin_macro): encapsulate plugindiagemitter
