+++ crates/swc_plugin_macro/src/lib.rs
@@ -32,6 +32,30 @@ fn handle_func(func: ItemFn) -> TokenStream {
         #[cfg(target_arch = "wasm32")] // Allow testing
         extern "C" {
             fn __set_transform_result(bytes_ptr: i32, bytes_ptr_len: i32);
+            fn __emit_diagnostics(bytes_ptr: i32, bytes_ptr_len: i32);
+            fn __free(bytes_ptr: i32, size: i32) -> i32;
+        }
+
+        /// An emitter for the Diagnostic in plugin's context by borrowing host's
+        /// environment context.
+        ///
+        /// It is not expected to call this directly inside of plugin transform.
+        /// Instead, it is encouraged to use global HANDLER.
+        pub struct PluginDiagnosticsEmitter;
+
+        impl swc_core::common::errors::Emitter for PluginDiagnosticsEmitter {
+            #[cfg_attr(not(target_arch = "wasm32"), allow(unused))]
+            fn emit(&mut self, db: &swc_core::common::errors::DiagnosticBuilder<'_>) {
+                let diagnostic = swc_core::common::plugin::serialized::VersionedSerializable::new((*db.diagnostic).clone());
+                let diag = swc_core::common::plugin::serialized::PluginSerializedBytes::try_serialize(&diagnostic)
+                    .expect("Should able to serialize Diagnostic");
+                let (ptr, len) = diag.as_ptr();
+
+                #[cfg(target_arch = "wasm32")] // Allow testing
+                unsafe {
+                    __emit_diagnostics(ptr as i32, len as i32);
+                }
+            }
         }
 
 
@@ -83,7 +107,7 @@ fn handle_func(func: ItemFn) -> TokenStream {
             let handler = swc_core::common::errors::Handler::with_emitter(
                 true,
                 false,
-                Box::new(swc_core::plugin::errors::PluginDiagnosticsEmitter)
+                Box::new(PluginDiagnosticsEmitter)
             );
             let handler_set_result = swc_core::plugin::errors::HANDLER.inner.set(handler);
